% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{create_connection_sf}
\alias{create_connection_sf}
\title{Create connection with snowflake}
\usage{
create_connection_sf()
}
\value{
Connection
}
\description{
Create connection with snowflake db. Remember to define the variables
\code{snowflake_uid}, \code{snowflake_password}, \code{snowflake_warehouse}, \code{snowflake_database}
and \code{snowflake_scheme} with the \code{keyring} package (\code{keyring::key_set()}).
}
