% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dens}
\alias{dens}
\title{Simple density plot function}
\usage{
dens(column, alpha = NA)
}
\arguments{
\item{column}{Name of the column}

\item{alpha}{Alpha parameter for ggplot}
}
\description{
Simple density plot function
}
