% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-env-var.R
\name{df_to_enviroment}
\alias{df_to_enviroment}
\title{Creates a enviroment with variables from a pairwise dataframe}
\usage{
df_to_enviroment(data)
}
\arguments{
\item{data}{A dataframe with two columns}
}
\value{
An enviroment
}
\description{
Creates a enviroment with variables from a pairwise dataframe
}
