% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{read_query}
\alias{read_query}
\title{Read the content of a text file}
\usage{
read_query(file)
}
\arguments{
\item{file}{Location of the file}
}
\value{
String
}
\description{
Read the content of a text file
}
