% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{run_queries}
\alias{run_queries}
\title{Run multiples queries in a single folder}
\usage{
run_queries(folder = "sql/", save = "data/")
}
\arguments{
\item{folder}{Folder with all the queries}

\item{save}{Folder to store the results of the queries}
}
\description{
Run multiples queries in a single folder
}
