% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{run_query}
\alias{run_query}
\title{Run query}
\usage{
run_query(query, file = NA, name = NA)
}
\arguments{
\item{query}{Query in a string format}

\item{file}{Route of the query}
}
\value{
Dataframe
}
\description{
Execute a query, can be a string or a query from a file.
}
